+++ NexusKrop.IceCube/Interop/LibC.cs
@@ -1,6 +1,4 @@
-﻿#if NET6_0_OR_GREATER
-
-namespace NexusKrop.IceCube.Interop;
+﻿namespace NexusKrop.IceCube.Interop;
 using System;
 using System.Collections.Generic;
 using System.Linq;
@@ -9,11 +7,11 @@ using System.Runtime.Versioning;
 using System.Text;
 using System.Threading.Tasks;
 
+#if NET6_0_OR_GREATER
 [SupportedOSPlatform("linux")]
+#endif
 internal static class LibC
 {
     [DllImport("libc.so.6", SetLastError = true)]
     internal static extern int kill(int pid, int signum);
 }
-
-#endif